+++ packages/nutui-inject-ui-styles/src/index.ts
@@ -1,0 +1,3 @@
+import { babelComponentStyle } from './babel-component-style'
+
+export default babelComponentStyle