+++ src/packages/avatarcropper/avatarcropper.taro.tsx
@@ -4,8 +4,9 @@ import React, {
   useMemo,
   useCallback,
   FunctionComponent,
+  useLayoutEffect,
 } from 'react'
-import Taro, { useReady, createSelectorQuery } from '@tarojs/taro'
+import Taro, { createSelectorQuery } from '@tarojs/taro'
 import classNames from 'classnames'
 import { Canvas, CommonEventFunction, View } from '@tarojs/components'
 import { getWindowInfo } from '@/utils/taro/get-system-info'
@@ -137,7 +138,7 @@ export const AvatarCropper: FunctionComponent<
     cropperCanvasContext: null,
   })
 
-  useReady(() => {
+  useLayoutEffect(() => {
     if (showAlipayCanvas2D) {
       const { canvasId } = canvasAll
       createSelectorQuery()
@@ -149,7 +150,7 @@ export const AvatarCropper: FunctionComponent<
         })
         .exec()
     }
-  })
+  }, [showAlipayCanvas2D, state.displayHeight, state.displayWidth])
 
   useEffect(() => {
     setCanvasAll({
@@ -693,4 +694,5 @@ export const AvatarCropper: FunctionComponent<
     </>
   )
 }
+
 AvatarCropper.displayName = 'NutAvatarCropper'