+++ src/packages/dialog/demos/h5/demo6.tsx
@@ -22,7 +22,14 @@ const Demo6 = () => {
           '--nutui-dialog-close-color': '#888B94',
         }}
         onConfirm={() => setVisible1(false)}
-        onCancel={() => setVisible1(false)}
+        onClose={() => {
+          setVisible1(false)
+          console.log('onClose')
+        }}
+        onCancel={() => {
+          // do sth else
+          console.log('onCancel')
+        }}
       >
         支持函数调用和组件调用两种方式。
       </Dialog>